backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic[email]==2.12.5
python-dotenv==1.0.1
httpx==0.28.1
requests==2.33.1

# Gemini SDK (new)
google-genai

# MongoDB async
motor

# Redis async
redis

# WebSockets (already bundled with fastapi/starlette but explicit)
websockets==16.0

# JWT for calendar auth
python-jose[cryptography]

# CalDAV (Apple Calendar)
caldav

# BSON for ObjectId
pymongo
